add general boardpage draft
