--- src/pages/MatchPage/MatchPage.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const MatchPage: React.FC = () => {
-  return <div>직관 매칭</div>;
-};
-
-export default MatchPage;
